% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cosine_dissimilarity}
\alias{cosine_dissimilarity}
\title{Compute Cosine Dissimilarity Matrix}
\usage{
cosine_dissimilarity(mat, root = FALSE)
}
\arguments{
\item{mat}{A numeric matrix or data frame.}

\item{root}{A logical value indicating whether to take the square root of the cosine dissimilarity
matrix. Default is FALSE.}
}
\value{
A dissimilarity matrix of class "dist" containing the cosine dissimilarity between each pair of rows.
}
\description{
This function computes the cosine dissimilarity matrix based on the input matrix.
The cosine dissimilarity between two vectors is defined as 1 minus the cosine similarity.
}
\examples{
TRUE
}
