% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_themes.R
\name{generate_themes}
\alias{generate_themes}
\title{Cluster pathway embeddings and generate themes}
\usage{
generate_themes(PAVER_result, hclust_method = "ward.D2", ...)
}
\arguments{
\item{PAVER_result}{A list containing the result of running prepare_data}

\item{hclust_method}{The agglomeration method to be used in hierarchical clustering. Default is "ward.D2".}

\item{...}{Additional arguments to be passed to dynamicTreeCut}
}
\value{
A list containing the updated PAVER_result
}
\description{
This function clusters pathway embeddings using cosine similarity and generates themes for the clusters.
}
\examples{
TRUE

}
