% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gsea_example}
\alias{gsea_example}
\title{An example of GSEA results}
\format{
A data frame with 1080 rows and 4 columns:
\describe{
\item{GOID}{Gene Ontology Term IDs}
\item{SCZvsCtrl}{NES of GO terms}
\item{SCZvsCtrlD}{NES of GO terms}
\item{SCZvsCtrlS}{NES of GO terms}
}
}
\source{
\url{https://doi.org/10.1038/s41380-021-01205-y}
}
\usage{
gsea_example
}
\description{
A wide data frame with Normalized Enrichment Scores from
Gene-Set Enrichment Analysis of three different comparisons.
}
\keyword{datasets}
