% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kegg_example}
\alias{kegg_example}
\title{An example of KEGG results}
\format{
A data frame with 183 rows and 2 columns:
\describe{
\item{MAP}{KEGG Pathway IDS}
\item{MA}{Enrichment Scores of KEGG pathways}
}
}
\source{
\url{https://doi.org/10.1101/2023.10.13.562226}
}
\usage{
kegg_example
}
\description{
A wide data frame with enrichment scores (log10 FDR) from
MetaboAnalyst of one comparison.
}
\keyword{datasets}
