% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAVER_combined_plot.R
\name{PAVER_combined_plot}
\alias{PAVER_combined_plot}
\title{PAVER_combined_plot}
\usage{
PAVER_combined_plot(PAVER_result, unit = NULL)
}
\arguments{
\item{PAVER_result}{a list containing the output of PAVER analysis}

\item{unit}{optionally, the unit of enrichment analysis for the figure legend title.}
}
\value{
A combined ggplot2 figure consisting of the PAVER_cluster_plot, PAVER_regulation_plot,
PAVER_interpretation_plot, and PAVER_hunter_plot.
}
\description{
This function takes a PAVER result object and generates a combined ggplot2
figure consisting of the PAVER_cluster_plot, PAVER_regulation_plot,
PAVER_interpretation_plot, and PAVER_hunter_plot.
}
\examples{
TRUE

}
