% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAVER_export.R
\name{PAVER_export}
\alias{PAVER_export}
\title{Export PAVER results to a formatted data frame}
\usage{
PAVER_export(PAVER_result)
}
\arguments{
\item{PAVER_result}{A PAVER result object generated by the \code{\link{generate_themes}} function.}
}
\value{
A wide format data frame with GO Term IDs, Definitions, gererated themes, and columns
for each of the groups in the PAVER result, with values representing the original input value
of each GO term for that group.
}
\description{
This function takes a PAVER result object and prepares it for export as a wide format
data frame. The resulting data frame will have GO Term IDs, Definitions, generated themes, and columns
for each of the groups in the PAVER result, with values representing the original input value
of each GO term for that group.
}
\examples{
TRUE

}
