% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAVER_hunter_plot.R
\name{PAVER_hunter_plot}
\alias{PAVER_hunter_plot}
\title{PAVER_hunter_plot}
\usage{
PAVER_hunter_plot(PAVER_result)
}
\arguments{
\item{PAVER_result}{a list containing the output of PAVER analysis}
}
\value{
A heatmap of the expression data with clustering and color-coded values based on the direction of regulation.
}
\description{
This function takes a PAVER result object and generates a heatmap of the enrichment analysis
with clustering and color-coded values based on the direction of regulation.
}
\examples{
TRUE

}
