% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAVER_interpretation_plot.R
\name{PAVER_interpretation_plot}
\alias{PAVER_interpretation_plot}
\title{PAVER_interpretation_plot}
\usage{
PAVER_interpretation_plot(PAVER_result)
}
\arguments{
\item{PAVER_result}{a list containing the output of PAVER analysis}
}
\value{
A ggplot2 scatterplot of the MDS layout with labeled points colored by group.
}
\description{
This function takes a PAVER result object and generates a ggplot2 scatterplot
of the MDS layout with labeled points colored by group.
}
\examples{
TRUE

}
