% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAVER_regulation_plot.R
\name{PAVER_regulation_plot}
\alias{PAVER_regulation_plot}
\title{PAVER_regulation_plot}
\usage{
PAVER_regulation_plot(PAVER_result)
}
\arguments{
\item{PAVER_result}{a list containing the output of PAVER analysis}
}
\value{
A ggplot2 scatterplot of the UMAP layout colored by direction of regulation (up or down) and faceted for each group.
}
\description{
This function takes a PAVER result object and generates a ggplot2 scatterplot
of the UMAP layout colored by direction of regulation (up or down) for each group.
}
\examples{
TRUE

}
