% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare pathway analysis results for downstream theme generation and summary visualization}
\usage{
prepare_data(input, embeddings, term2name)
}
\arguments{
\item{input}{A data frame in wide format containing the pathway analysis results. The first column is expected to be term IDs, e.g. GO:0000001, while the remaining columns are expected to be numeric values representing the enrichment of each term in each group.}

\item{embeddings}{A matrix containing the precomputed embeddings of the input terms.}

\item{term2name}{A data frame containing two columns that map pathway term IDs to pathway term names.}
}
\value{
A PAVER_result list with the prepared data ready for theme generation
}
\description{
This function prepares the results of pathway analysis for downstream theme generation and visualization using precomputed embeddings.
}
\examples{
TRUE

}
