% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAVER_theme_plot.R
\name{PAVER_theme_plot}
\alias{PAVER_theme_plot}
\title{PAVER_theme_plot}
\usage{
PAVER_theme_plot(PAVER_result, show_cluster_legend = TRUE)
}
\arguments{
\item{PAVER_result}{a list containing the output of PAVER analysis}

\item{show_cluster_legend}{a logical indicating whether to show the cluster legend.}
}
\value{
A ggplot2 scatterplot of the UMAP layout colored by cluster assignments
}
\description{
This function takes a PAVER result object and generates a ggplot2 scatterplot
of the UMAP layout colored by cluster assignments.
}
\examples{
TRUE

}
