% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare Gene Ontology enrichment analysis results for downstream theme generation and summary visualization}
\usage{
prepare_data(input, embeddings, ontology_index, unit = "NES", seed = 123)
}
\arguments{
\item{input}{A data frame in wide format containing the Gene Ontology enrichment analysis results. The first column is expected to be GO IDs, e.g., GO:0000001}

\item{embeddings}{A matrix containing the precomputed Gene Ontology embeddings}

\item{ontology_index}{An ontology_index object prepared from the core Gene Ontology used to compute embeddings}

\item{unit}{A character describing the unit of the enrichment analysis results}

\item{seed}{An integer setting the seed for UMAP}
}
\value{
A PAVER_result list with the prepared data, embedding matrix, GO Term ID to definition mappings, and computed UMAP
}
\description{
This function prepares the results of Gene Ontology enrichment analysis for downstream theme generation and visualization by using precomputed embeddings of the core Gene Ontology.
}
\examples{
TRUE

}
